process/requirements.txt
lxml
isodate
psycopg2-binary
pandas
tqdm
openpyxl